last commit

+++ main.py
@@ -73,6 +73,7 @@ class Request:
 
     def __init__(self, request_type):
         Request.id += 1
+        self.wait = 0
         self.enter_queue_time = 0
         self.exit_queue_time = 0
         self.id = Request.id
@@ -126,6 +127,7 @@ class Request:
 
 def check_error(service_type):
     p = 0
+
     if service_type == "mobile portal service":
         p = 0.01
     elif service_type == "order management service":
@@ -171,52 +173,46 @@ def request_generator(rate):
 
 def general_service(request):
     global env
+    global timeout_fails
     service = request.current_process()
     service.add_customer_to_queue(request)
     request.enter_queue_time = env.now
-    # print(service.type, "time_entered_queue: ", request.enter_queue_time, request.id)
     with service.get_resources().request(priority=request.priority) as req:
         yield req
-        # service_queue = service.get_queue()
-        # preferred_index = find_preferred_request(service_queue)
-        # request = service_queue[preferred_index]
-        # if len(service_queue) > 0:
         request.exit_queue_time = env.now
-        service_start_time = env.now
-        service.remove_customer_from_queue(request)
-        service.add_in_progress_request(request, env.now)
-        # print(service.type, "time_left_queue: ", request.exit_queue_time, request.id)
         time_in_queue = request.exit_queue_time - request.enter_queue_time
         request.add_waiting_time(time_in_queue)
+        request.wait += time_in_queue
         service.add_time_to_queue_time_sum(time_in_queue)
-        if len(request.processes) > request.turn + 1:
-            request.next_process()
-            yield env.process(general_service(request))
-            request.turn -= 1
-        service_time = random.expovariate(service.mean)
-        # print(service.type, "service time: ", service_time)
-        yield env.timeout(service_time)
-        service.add_busy_time(env.now - service_start_time)
-        service.remove_in_progress_requests(request)
-        # print(service.type, "service finished in", env.now)
-
-
-# def find_preferred_request(queue):
-#     try:
-#         preferred_request = queue[0]
-#         index = 0
-#         for i in range(1, len(queue)):
-#             request = queue[i]
-#             if request.priority < preferred_request.priority:
-#                 preferred_request = request
-#                 index = i
-#         return index
-#     except:
-#         print("Error occurred in find prefer request function!")
-#         return 0
+        check = check_error(service_type=service.type)
+        if check:
+            request.exit_queue_time = env.now
+            service.remove_customer_from_queue(request)
+            service.add_busy_time(0)
+            service.remove_in_progress_requests(request)
+            error_fails[request.type - 1] += 1
+        if request.wait >= maximum_waiting_times[request.type - 1]:
+            service.remove_customer_from_queue(request)
+            service.add_busy_time(0)
+            service.remove_in_progress_requests(request)
+            timeout_fails[request.type - 1] += 1
+        else:
+            service_start_time = env.now
+            service.remove_customer_from_queue(request)
+            service.add_in_progress_request(request, env.now)
+            if len(request.processes) > request.turn + 1:
+                request.next_process()
+                yield env.process(general_service(request))
+                request.turn -= 1
+            service_time = random.expovariate(service.mean)
+            yield env.timeout(service_time)
+            service.add_busy_time(env.now - service_start_time)
+            service.remove_in_progress_requests(request)
 
 
 env = simpy.Environment()
+error_fails = [0 for _ in range(7)]
+timeout_fails = [0 for _ in range(7)]
 
 number_of_resources = list(map(int, input().split()))
 
@@ -298,4 +294,44 @@ print("     Delivery communication:",
       Service.Servers_busy_times[5] / (simulation_time * number_of_resources[3]))
 print("     Restaurant management service:",
       Service.Servers_busy_times[6] / (simulation_time * number_of_resources[0]))
+print('-' * 50)
+
+print('-' * 50)
+print("The percentage of error for each requestt:")
+print("     Place order through mobile phone:",
+      error_fails[0] / Request.number_of_received_requests[0])
+print("     Place order through the web:",
+      error_fails[1] / Request.number_of_received_requests[1])
+print("     Send a message to the courier:",
+      error_fails[2] / Request.number_of_received_requests[2])
+print("     See restaurant information via mobile:",
+      error_fails[3] / Request.number_of_received_requests[3])
+print("     View restaurant information through the web:",
+      error_fails[4] / Request.number_of_received_requests[4])
+print("     Courier request:",
+      error_fails[5] / Request.number_of_received_requests[5])
+print("     Order tracking:",
+      error_fails[6] / Request.number_of_received_requests[6])
+print("     Timeout percentage in total:",
+      sum(error_fails) / sum(Request.number_of_received_requests))
+print('-' * 50)
+
+print('-' * 50)
+print("percentage of timeout for requests:")
+print("     Place order through mobile phone:",
+      timeout_fails[0] / Request.number_of_received_requests[0])
+print("     Place order through the web:",
+      timeout_fails[1] / Request.number_of_received_requests[1])
+print("     Send a message to the courier:",
+      timeout_fails[2] / Request.number_of_received_requests[2])
+print("     See restaurant information via mobile:",
+      timeout_fails[3] / Request.number_of_received_requests[3])
+print("     View restaurant information through the web:",
+      timeout_fails[4] / Request.number_of_received_requests[4])
+print("     Courier request:",
+      timeout_fails[5] / Request.number_of_received_requests[5])
+print("     Order tracking:",
+      timeout_fails[6] / Request.number_of_received_requests[6])
+print("     Timeout percentage in total:",
+      sum(timeout_fails) / sum(Request.number_of_received_requests))
 print('-' * 50)
